--- src/fmcore/types/enums/aws_enums.py
@@ -1,14 +1,0 @@
-from enum import Enum
-
-
-class AWSRegion(str, Enum):
-    US_EAST_1 = "us-east-1"
-    US_EAST_2 = "us-east-2"
-    US_WEST_1 = "us-west-1"
-    US_WEST_2 = "us-west-2"
-    EU_CENTRAL_1 = "eu-central-1"
-    EU_WEST_1 = "eu-west-1"
-    AP_SOUTHEAST_1 = "ap-southeast-1"
-    AP_SOUTHEAST_2 = "ap-southeast-2"
-    AP_NORTHEAST_1 = "ap-northeast-1"
-    SA_EAST_1 = "sa-east-1"
