--- src/fmcore/types/enums/provider_enums.py
@@ -1,9 +1,0 @@
-from enum import Enum
-
-
-class ProviderType(str, Enum):
-    """Enum representing different provider types."""
-
-    BEDROCK = "BEDROCK"
-    LAMBDA = "LAMBDA"
-    OPENAI = "OPENAI"
